+++ examples/api/bsoncxx/examples/validation/validator.cpp
@@ -77,8 +77,8 @@ void example(std::uint8_t const* bytes, std::size_t length) {
 
         EXPECT(!bsoncxx::validate(bytes, length, options, &offset));
 
-        // Offset of `"$numberInt": "123"` relative to start of the sub-document. (CDRIVER-5710)
-        EXPECT(offset == 4u);
+        // Offset of `"$numberInt": "123"` relative to start of the document.
+        EXPECT(offset == 31u);
     }
 }
 // [Example]