+++ CHANGELOG.md
@@ -9,7 +9,9 @@ Changes prior to 3.9.0 are documented as [release notes on GitHub](https://githu
 
 ## 4.1.3 [Unreleased]
 
-<!-- Will contain entries for the next patch release. -->
+### Changed
+
+- Bump the auto-downloaded C Driver version to [2.0.2](https://github.com/mongodb/mongo-c-driver/releases/tag/2.0.2).
 
 ## 4.1.2